Remove `<iostream>` includes from non-test files
Most includes of `<iostream>` can be replaced with `<ostream>` since the
only usage is an ostream overload or print operation. In `hash_map.hpp`
the include isn't needed at all and is removed.

This did require a signature change for `print()` to `print(std::ostream& o)`
in `strod.hpp` so that the print function is using a generic `std::ostream`
and not `std::cout` directly.